+++ ui/App/views/Mods/components/ModList.jsx
@@ -1,5 +1,6 @@
 import Mod from "./Mod";
 import React from "react";
+import {t} from "../../../../identity/preferences";
 
 
 const ModList = ({mods, factorioVersion, updateMod, toggleMod, deleteMod, addUpdatableMod = null, disabled = false}) => {
@@ -8,11 +9,12 @@ const ModList = ({mods, factorioVersion, updateMod, toggleMod, deleteMod, addUpd
         <table className="w-full">
             <thead>
             <tr className="text-left py-1">
-                <th>Name</th>
-                <th>Enabled</th>
-                <th>Compatibility</th>
-                <th>Mod Version</th>
-                <th>Factorio Version</th>
+                <th>{t('mods.name')}</th>
+                <th>{t('mods.enabled')}</th>
+                <th>{t('mods.compatibility')}</th>
+                <th>{t('mods.dependencies')}</th>
+                <th>{t('mods.modVersion')}</th>
+                <th>{t('mods.factorioVersion')}</th>
                 <th/>
             </tr>
             </thead>